+++ __tests__/utils/CustomErrorTest.js
@@ -1,4 +1,4 @@
-import CustomError from '../../src/utils/CustomError';
+import CustomError from '../../src/exceptions/CustomError.js';
 
 describe('CustomError 클래스 테스트', () => {
   test('에러 헤더 [ERROR]를 포함한 메시지를 출력한다.', () => {
