+++ src/User.js
@@ -1,7 +1,7 @@
 import DongHang from './DongHang.js';
 import ERROR_MESSAGE from './constants/error.js';
 import { LOTTO_PRICE } from './constants/number.js';
-import CustomError from './customs/CustomError.js';
+import CustomError from './exceptions/CustomError.js';
 import Input from './utils/Input.js';
 import PROMPT from './constants/prompt.js';
 import NumberValidator from './validators/NumberValidator.js';
