+++ src/exceptions/CustomError.js
